First release.
This is the first release of LuaCue 1.0.0. It has decoding
functionality, ready to be used and abused. Thanks to the
"Wikipedia article on cuesheets":[1] as provided an
easy-to-understand documentation of the cuesheet format.
Also, thanks to "lhf":[2] and "jA_cOp":[3] for helping me out in
the Lua IRC channel.
List of features:

  * Cuesheet decoding.
  * Basic UTF-8 compliance (it "will work" with UTF-8 input files)

[1]:http://en.wikipedia.org/wiki/Cue_sheet_(computing)
[2]:http://www.tecgraf.puc-rio.br/~lhf/
[3]:http://github.com/JakobOvrum